MCU change, again
STM32L0 had quite a big problem, common reference for the comparators.
Another minus was the missing SWO for debugging.

STM32L4 has floating point, and most importantly, has embedded OPAMPS!
This means I can directly plug in the filter to the analog inputs and
everything is software. 2 opamps and 2 comparators (if needed).
External circuitry would be minimal for the PIR sensor.

It also means that I can read with the embedded 12-bit ADC the
amplified values. Precision is not crucial, we are interested in dips
and swells.