Sensor logic implemented using MCU opamps
Of the ST PIR design I use the first 2 stages only, so no comparators.
1st stage using OPAMP1
2nd stage using OPAMP2

OPAMP1 amplifies the sensor output around the DC component of the PIR
sensor.

OPAMP2 amplifies the stage 1 output around half of VCC.

The output is then passed differentially to ADC which gives a
difference between stage 2 output and VCC/2.

This can be later converted into real volts by using the internal
VREFINT of the MCU because VCC can be measured against it. ST has
some (hopefully) cool factory calibration feature to compensate
for temperature deviations of the ADC which makes all this
possible  without external sigma-delta ADC.

The ADC has analog window thresholds that can be used to trigger
interrupts when input is beyond threshold. This means it is
possible to set sensitivity by modifying this threshold instead
of playing with the OPAMP2 gain (which anyway has only 4 levels).

(Just wasted a whole fucking saturday on this...)