KEY_OK is now SYS_WKUP5, timebase changed to TIM1
Regenerated all code. As in this commit, the code can be regene-
rated using the hal.ioc file loaded up in STM32CubeMX.

This will add additional files, or modify existing ones. I need
to make sure to put code under USER CODE and not otherwise.

Makefile is heavily reworked due to bugs in CubeMX generation so
that one has to be manually merged.