Polishing of the camera interface + settings
Camera interface settings are stored in the *settings* module and
all that goes into/out of it gets through some functions. For now
the functions are basic but semaphores will be added later.

This commit also adds intervalometer style parameters to the
trigger settings and can now accomodate for all sort of inter-delays
in-between and repetition of triggers.