Implemented input -> core ipc queue
I can now pass info about what key was pressed between the tInput
and tCore tasks.

On problem I encountered was the fact that FreeRTOS blocks the
usage of some of its ISR API if the NVIC Priority is above a certain
threshold. In my case the threshold is 5 so I just modified the
priority of the EXTI interrupt to 5 to allow that API subset to
function properly.

The queues are currently implemented using dynamic allocation and
I fucking hate that, doesn't make sense to allocate/deallocate
every time I send an 8 byte message. Need to investigate how good
(if any) is the support of cmsis_os for FreeRTOS static allocations.